+++ rgwin_pair.c
@@ -48,7 +48,7 @@ void win_pair_rearrange(window_t *win, grect_t *box, data_metrics_t *metrics)
 {
     window_pair_t *dwin = win->data;
     grect_t box1, box2;
-    int min, diff, split, splitwid, max;
+    double min, diff, split, splitwid, max;
     window_t *key;
     window_t *ch1, *ch2;
 
@@ -135,7 +135,7 @@ void win_pair_rearrange(window_t *win, grect_t *box, data_metrics_t *metrics)
 
     if (dwin->vertical) {
         dwin->splitpos = split;
-        dwin->splitwidth = splitwid;
+        dwin->splitwidth = (int) splitwid;
         box1.left = win->bbox.left;
         box1.right = dwin->splitpos;
         box2.left = box1.right + dwin->splitwidth;
@@ -155,7 +155,7 @@ void win_pair_rearrange(window_t *win, grect_t *box, data_metrics_t *metrics)
     }
     else {
         dwin->splitpos = split;
-        dwin->splitwidth = splitwid;
+        dwin->splitwidth = (int) splitwid;
         box1.top = win->bbox.top;
         box1.bottom = dwin->splitpos;
         box2.top = box1.bottom + dwin->splitwidth;
